+++ mrmustard/lab/gates.py
@@ -430,13 +430,63 @@ class S2gate(Parametrized, Transformation):
 class Interferometer(Parametrized, Transformation):
     r"""N-mode interferometer.
 
-    It corresponds to a Ggate with zero mean and a ``2N x 2N`` orthogonal symplectic matrix.
+    It corresponds to a Ggate with zero mean and a ``2N x 2N`` unitary symplectic matrix.
 
     Args:
         num_modes (int): the num_modes-mode interferometer
-        orthogonal (2d array): a valid orthogonal matrix. For N modes it must have shape `(2N,2N)`
-        orthogonal_trainable (bool): whether orthogonal is a trainable variable
+        unitary (2d array): a valid unitary matrix U. For N modes it must have shape `(N,N)`
+        unitary_trainable (bool): whether unitary is a trainable variable
         modes (optional, List[int]): the list of modes this gate is applied to
+        ortho_symplectic (2d array): a valid orthogonal symplectic matrix that comes from its unitary matrix U=X+iY, and [[X, Y],[-Y, X]]. For N modes it must have shape `(2N,2N)`
+    """
+
+    def __init__(
+        self,
+        num_modes: int,
+        unitary: Optional[Tensor] = None,
+        unitary_trainable: bool = False,
+        modes: Optional[List[int]] = None,
+    ):
+        if modes is not None and num_modes != len(modes):
+            raise ValueError("Invalid number of modes and the mode list here!")
+        if unitary is None:
+            unitary = math.random_unitary(num_modes)
+        super().__init__(
+            unitary=unitary,
+            unitary_trainable=unitary_trainable,
+        )
+        self._modes = modes or list(range(num_modes))
+        self.is_gaussian = True
+
+    @property
+    def X_matrix(self):
+        return math.block(
+            [
+                [math.real(self.unitary.value), -math.imag(self.unitary.value)],
+                [math.imag(self.unitary.value), math.real(self.unitary.value)],
+            ]
+        )
+
+    def _validate_modes(self, modes):
+        if len(modes) != self.unitary.value.shape[-1]:
+            raise ValueError(
+                f"Invalid number of modes: {len(modes)} (should be {self.unitary.shape[-1]})"
+            )
+
+    def __repr__(self):
+        modes = self.modes
+        unitary = repr(math.asnumpy(self.unitary.value)).replace("\n", "")
+        return f"Interferometer(num_modes = {len(modes)}, unitary = {unitary}){modes}"
+
+
+class RealInterferometer(Parametrized, Transformation):
+    r"""N-mode interferometer parametrized by an NxN orthogonal matrix (or 2N x 2N block-diagonal orthogonal matrix). This interferometer does not mix q and p.
+    Does not mix q's and p's.
+
+    Args:
+        orthogonal (2d array, optional): a real unitary (orthogonal) matrix. For N modes it must have shape `(N,N)`.
+            If set to `None` a random real unitary (orthogonal) matrix is used.
+        orthogonal_trainable (bool): whether orthogonal is a trainable variable
     """
 
     def __init__(
@@ -446,63 +496,19 @@ class Interferometer(Parametrized, Transformation):
         orthogonal_trainable: bool = False,
         modes: Optional[List[int]] = None,
     ):
-        if modes is not None and (
-            num_modes != len(modes) or any(mode >= num_modes for mode in modes)
-        ):
+        if modes is not None and num_modes != len(modes):
             raise ValueError("Invalid number of modes and the mode list here!")
-        if orthogonal is None:
-            U = math.random_unitary(num_modes)
-            orthogonal = math.block([[math.real(U), -math.imag(U)], [math.imag(U), math.real(U)]])
-        super().__init__(
-            orthogonal=orthogonal,
-            orthogonal_trainable=orthogonal_trainable,
-        )
-        self._modes = modes or list(range(num_modes))
-        self.is_gaussian = True
-
-    @property
-    def X_matrix(self):
-        return self.orthogonal.value
-
-    def _validate_modes(self, modes):
-        if len(modes) != self.orthogonal.value.shape[-1] // 2:
-            raise ValueError(
-                f"Invalid number of modes: {len(modes)} (should be {self.orthogonal.shape[-1] // 2})"
-            )
-
-    def __repr__(self):
-        modes = self.modes
-        orthogonal = repr(math.asnumpy(self.orthogonal.value)).replace("\n", "")
-        return f"Interferometer(num_modes = {len(modes)}, orthogonal = {orthogonal}){modes}"
-
-
-class RealInterferometer(Parametrized, Transformation):
-    r"""N-mode interferometer with a real unitary matrix (or block-diagonal orthogonal matrix).
-    Does not mix q's and p's.
-
-    Args:
-        orthogonal (2d array, optional): a valid orthogonal matrix. For N modes it must have shape `(N,N)`.
-            If set to `None` a random orthogonal matrix is used.
-        orthogonal_trainable (bool): whether orthogonal is a trainable variable
-    """
-
-    def __init__(
-        self,
-        num_modes: int,
-        orthogonal: Optional[Tensor] = None,
-        orthogonal_trainable: bool = False,
-    ):
         if orthogonal is None:
             orthogonal = math.random_orthogonal(num_modes)
         super().__init__(orthogonal=orthogonal, orthogonal_trainable=orthogonal_trainable)
-        self._modes = list(range(num_modes))
+        self._modes = modes or list(range(num_modes))
         self._is_gaussian = True
 
     @property
     def X_matrix(self):
         return math.block(
             [
-                [self.orthogonal.value, math.zeros_like(self.orthogonal.value)],
+                [self.orthogonal.value, -math.zeros_like(self.orthogonal.value)],
                 [math.zeros_like(self.orthogonal.value), self.orthogonal.value],
             ]
         )
@@ -564,6 +570,7 @@ class Ggate(Parametrized, Transformation):
 # ~~~~~~~~~~~~~
 # NON-UNITARY
 # ~~~~~~~~~~~~~
+
 
 # pylint: disable=no-member
 class Attenuator(Parametrized, Transformation):