+++ mrmustard/math/__init__.py
@@ -49,6 +49,7 @@ class Math:
     r"""
     This class is a switcher for performing math operations on the currently active backend.
     """
+
     # pylint: disable=no-else-return
     def __getattribute__(self, name):
         if settings.backend == "tensorflow":